Clean up TensorToPlasma, fix a rare Stream cleanup bug
Previously running the test py file will result in CUDA_ERROR_DEINITIALIZED.  I
think this is because the custom Stream is double-freed.  This is now fixed by
changing it from a shared_ptr to a raw pointer and manually delete.

(The bug was not manifested in running the full sgd example, though.)

Another change: change the sync method from ThenWaitFor(Event) to
ThenWaitFor(Stream).  This has the same performance, and reduces code.